+++ frontend/index.html
@@ -31,7 +31,6 @@
       }
     </style>
     <div id="root"></div>
-    <!-- TODO: fichier semble chargé deux fois -->
     <script type="module" src="./src/App.tsx"></script>
   </body>
 </html>
