+++ frontend/src/common/types/book.enum.ts
