+++ frontend/src/common/types/graphic-asset-type.enum.ts
