+++ frontend/src/common/types/graphic-asset.entity.ts
