+++ frontend/src/common/types/page.ts
