+++ frontend/src/module/book/canvas/element/DrawableElement.ts
@@ -1,5 +1,5 @@
 import * as fabric from 'fabric';
-import { Element } from '@/types/book';
+import { Element } from '@apptypes/book';
 export interface DrawableElement {
   draw(canvas: fabric.Canvas): void;
   update(attrs: Element): void;
