+++ frontend/src/module/book/canvas/element/types/SVG.ts
@@ -1,6 +1,6 @@
 import { DrawableElement } from '../DrawableElement'; // Adjust the path as needed
 import * as fabric from 'fabric';
-import { SVGElement } from '@/types/book';
+import { SVGElement } from '@apptypes/book';
 import { getAPIURL } from '@/common/utils/api';
 
 export class SVG implements DrawableElement {
