+++ frontend/src/module/book/canvas/hooks/usePageCreation.tsx
@@ -1,4 +1,4 @@
-import { Page } from '@/types/book';
+import { Page } from '@apptypes/book';
 import * as fabric from 'fabric';
 import { useEffect, useState } from 'react';
 import { ElementFactory } from '../element/ElementFactory';
