+++ frontend/src/module/book/components/BookHeader.tsx
@@ -2,7 +2,7 @@ import BreadCrumb from '@components/BreadCrumb';
 import Header from '@components/Header';
 import InlineEdit from '@components/InlineEdit';
 import { useAppSelector } from '@/common/hooks/useRedux';
-import { Book } from '@/types/book';
+import { Book } from '@apptypes/book';
 import { selectIsLoading } from '../Book.slice';
 
 const BookHeader: React.FC<{
