+++ frontend/src/module/book/sidePanel/graphicAssets/GraphicAssets.slice.ts
@@ -2,7 +2,7 @@ import { createSlice } from '@reduxjs/toolkit';
 
 import * as graphicAssetsActions from './GraphicAssets.actions.ts';
 import { RootState } from '@/common/store';
-import { GraphicAsset } from '@/types/graphic-asset.entity';
+import { GraphicAsset } from '@apptypes/graphic-asset.entity';
 
 export enum Mode {
   Default = 'DEFAULT',
