+++ frontend/tsconfig.app.json
@@ -14,8 +14,8 @@
       "@components/*": ["src/common/components/*"],
       "@services/*": ["src/services/*"],
       "@utils/*": ["src/common/utils/*"],
-      "@types/*": ["src/common/types/*"],
-      "@shared/*": ["src/shared/*"]
+      "@apptypes/*": ["src/common/types/*"],
+      "@shared/*": ["src/common/shared/*"]
     },
 
     /* Bundler mode */
